Update directory listings of meal-related fragments
