+++ app/src/main/java/com/example/checnutritionapp/ui/meal/ingredients/MealIngredientsViewModel.java
@@ -1,4 +1,4 @@
-package com.example.checnutritionapp;
+package com.example.checnutritionapp.ui.meal.ingredients;
 
 import androidx.lifecycle.ViewModel;
