+++ app/src/main/res/layout/meal_allergens_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ffffff"
-    tools:context=".MealAllergensFragment">
+    tools:context=".ui.meal.allergens.MealAllergensFragment">
 
     <TextView
         android:layout_width="wrap_content"
