+++ app/src/main/res/layout/meal_description_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MealDescription">
+    tools:context=".ui.meal.description.MealDescription">
 
     <TextView
         android:id="@+id/mealDescText"
