+++ app/src/main/res/layout/meal_fragment.xml
@@ -50,7 +50,7 @@
 
     <fragment
         android:id="@+id/fragment2"
-        android:name="com.example.checnutritionapp.MealDescription"
+        android:name="com.example.checnutritionapp.ui.meal.description.MealDescription"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="177dp"
@@ -66,7 +66,7 @@
 
     <fragment
         android:id="@+id/fragment3"
-        android:name="com.example.checnutritionapp.MealIngredients"
+        android:name="com.example.checnutritionapp.ui.meal.ingredients.MealIngredients"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         tools:layout_editor_absoluteX="17dp"
