+++ app/src/main/res/layout/meal_ingredients_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MealIngredients">
+    tools:context=".ui.meal.ingredients.MealIngredients">
 
     <TextView
         android:id="@+id/textView3"
