+++ app/src/main/res/layout/meal_nutrition_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ffffff"
-    tools:context=".MealNutrition">
+    tools:context=".ui.meal.nutrition.MealNutrition">
 
     <TextView
         android:id="@+id/nut0"
