+++ app/src/main/res/layout/meal_procedure_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ffffff"
-    tools:context=".MealProcedureFragment">
+    tools:context=".ui.meal.procedure.MealProcedureFragment">
 
     <TextView
         android:layout_width="wrap_content"
